+++ app/javascript/map.js
@@ -1,178 +1,153 @@
-// 位置情報取得
-document.addEventListener("turbo:load", function() {
-    const locationLink = document.getElementById('find-location-link');
-    
-    if (locationLink) {  // 要素が存在するか確認
-      locationLink.addEventListener('click', function(event) {
-        event.preventDefault(); // デフォルトのリンク動作をキャンセル
-        getLocationAndSearch();
-      });
-    }
-  });
-
-  function getLocationAndSearch() {
-    if (navigator.geolocation) {
-      navigator.geolocation.getCurrentPosition(function(position) {
-        const lat = position.coords.latitude;
-        const lng = position.coords.longitude;
-
-        console.log("Latitude: " + lat + ", Longitude: " + lng); // 緯度と経度をコンソールに出力
-
-        const searchUrl = document.getElementById('find-location-link').getAttribute('data-search-url');
-        const fullUrl = `${searchUrl}?lat=${lat}&lng=${lng}&from_current_location=true`;
-
-        window.location.href = fullUrl;
-      }, function(error) {
-        console.error("位置情報の取得に失敗しました: ", error);
-        alert("位置情報を取得できませんでした。東京を中心に表示します。");
-
-        // 位置情報がブロックされた場合、デフォルトの東京を表示するためにページを遷移
-        const searchUrl = document.getElementById('find-location-link').getAttribute('data-search-url');
-        const fullUrl = `${searchUrl}?lat=35.6895&lng=139.6917&from_current_location=false`;  // 東京の緯度経度をセット
-        window.location.href = fullUrl;  // 東京を中心にしたURLへ遷移
-      }
-    );
-  } else {
-    alert("ブラウザが位置情報取得に対応していません。");
-    // ブラウザが対応していない場合も東京を中心に表示
-    const searchUrl = document.getElementById('find-location-link').getAttribute('data-search-url');
-    const fullUrl = `${searchUrl}?lat=35.6895&lng=139.6917&from_current_location=false`;  // 東京の緯度経度をセット
-    window.location.href = fullUrl;  // 東京を中心にしたURLへ遷移
-  }
-}
-
 // 位置情報map表示
 const RamenMap = {
-    mapInitialized: false,
-    map: null,
-  
-    init: function(position) {
-      const userLocation = {
-        lat: position.coords.latitude,
-        lng: position.coords.longitude
-      };
-  
-      this.map = new google.maps.Map(document.getElementById("user-map"), {
-        zoom: 15,
-        center: userLocation
-      });
-  
-      this.createUserMarker(userLocation, this.map);
-  
-      // 地図の移動後に周辺のラーメン店を自動検索
-      this.addIdleListener();
-      
-      // 初回の店舗検索
-      this.searchNearby();
-    },
-  
-    createUserMarker: function(location, map) {
-      const marker = new google.maps.Marker({
-        position: location,
-        map: map,
-        title: "あなたの現在地"
-      });
-    },
-  
-    createMarker: function(place, map) {
-      const marker = new google.maps.Marker({
-        position: place.geometry.location,
-        map: map,
-        title: place.name,
-        icon: {
-          url: ramenIconPath,
-          scaledSize: new google.maps.Size(40, 40)
-        }
-      });
-  
-      const infoWindow = new google.maps.InfoWindow({
-        content: `<div><strong>${place.name}</strong></div>`
-      });
-  
-      marker.addListener('click', function() {
-        infoWindow.open(map, marker);
-      });
-    },
-  
-    searchNearby: function() {
-      if (!this.map) return;
-  
-      const center = this.map.getCenter();  // マップの中心を取得
-  
-      const request = {
-        location: center,
-        radius: '2000',
-        type: ['restaurant'],
-        keyword: 'らーめん 油そば'
-      };
-  
-      const service = new google.maps.places.PlacesService(this.map);
-      const self = this;
-  
-      service.nearbySearch(request, (results, status, pagination) => {
-        if (status === google.maps.places.PlacesServiceStatus.OK) {
-          for (let i = 0; i < results.length; i++) {
-            const place = results[i];
-            self.createMarker(place, self.map);
-          }
-  
-          // 次のページの結果がある場合
-          if (pagination && pagination.hasNextPage) {
-            setTimeout(() => {
-              pagination.nextPage();
-            }, 2000);
-          }
-        } else {
-          console.error("Places APIエラー: ", status);
-        }
-      });
-    },
-  
-    initDefaultMap: function() {
-      const tokyo = { lat: 35.6895, lng: 139.6917 };
-      this.map = new google.maps.Map(document.getElementById("user-map"), {
-        zoom: 15,
-        center: tokyo
-      });
-  
-      // 地図の移動後に周辺のラーメン店を自動検索
-      this.addIdleListener();
-  
-      // 初回の店舗検索
-      this.searchNearby();
+  map: null,
 
-      // 初期化済みとしてフラグを設定
-      this.mapInitialized = true;
+  // 共通の初期化処理
+  createMap: function(latitude, longitude, mapElementId, zoom = 15) {
+    const location = {
+      lat: latitude,
+      lng: longitude
+    };
+    this.map = new google.maps.Map(document.getElementById(mapElementId), {
+      zoom: zoom,
+      center: location
+    });
 
-    },
-  
-    // 地図が静止した後（ユーザーが地図を移動やズームした後）に店舗検索を実行
-    addIdleListener: function() {
-      const self = this;
-      this.map.addListener('idle', function() {
-        self.searchNearby();  // 地図の表示範囲が変更された後に検索
-      });
-    }
-  };
-  
-  // ページロード後にマップを初期化
-  document.addEventListener("turbo:load", function() {
-    if (document.getElementById("user-map") && !RamenMap.mapInitialized) {
-      if (navigator.geolocation) {
-        // 位置情報の取得を試みる
-        navigator.geolocation.getCurrentPosition(
-          function(position) {
-            // 位置情報の取得に成功した場合、ユーザーの位置を使用
-            RamenMap.init(position);
-          },
-          function() {
-            // 位置情報の取得に失敗した場合、東京を中心にしたマップを表示
-            RamenMap.initDefaultMap();
-          }
-        );
-      } else {
-        // ブラウザが位置情報取得に対応していない場合も東京を表示
-        RamenMap.initDefaultMap();
+    return this.map;
+  },
+
+  // location-map 用の初期化
+  initLocationMap: function(latitude, longitude) {
+    const map = this.createMap(latitude, longitude, "location-map");
+    // ラーメン店のマーカーを追加
+    this.searchNearby(map);
+    // マップが移動やズームされた後に再検索
+    this.addIdleListener(map);
+  },
+
+  // user-map 用の初期化
+  initUserMap: function(position) {
+    const latitude = position.coords.latitude;
+    const longitude = position.coords.longitude;
+    const map = this.createMap(latitude, longitude, "user-map");
+
+    // ユーザーの位置にマーカーを表示
+    this.addUserMarker(map, latitude, longitude);
+    this.searchNearby(map); // 周辺のラーメン店を検索してマーカーを表示
+    this.addIdleListener(map); // マップ移動後の再検索を追加
+  },
+
+  // store-map 用の初期化
+  initStoreMap: function(latitude, longitude) {
+    const map = this.createMap(latitude, longitude, "store-map");
+    // 店舗のマーカーを追加
+    this.addStoreMarker(map, latitude, longitude);
+  },
+
+  // マーカーを作成してマップに追加する
+  createMarker: function(place, map) {
+    const marker = new google.maps.Marker({
+      position: place.geometry.location,
+      map: map,
+      title: place.name,
+      icon: {
+        url: ramenIconPath, // ラーメン店の独自アイコン
+        scaledSize: new google.maps.Size(40, 40)
       }
+    });
+  
+    // 店舗名をクリックした際に遷移するリンクを含む InfoWindow を作成
+    const infoWindow = new google.maps.InfoWindow({
+      content: `<div><strong><a href="/ramen_shops/${place.place_id}">${place.name}</a></strong><br>${place.vicinity || ''}</div>` // 店舗名や住所を表示
+    });
+  
+    // マーカーがクリックされた時に InfoWindow を表示
+    marker.addListener('click', function() {
+      infoWindow.open(map, marker);
+    });
+  },
+
+  // ユーザーの位置にマーカーを追加
+  addUserMarker: function(map, latitude, longitude) {
+    new google.maps.Marker({
+      position: { lat: latitude, lng: longitude },
+      map: map,
+      title: "あなたの現在地"
+    });
+  },
+
+  // 店舗にマーカーを追加
+  addStoreMarker: function(map, latitude, longitude) {
+    new google.maps.Marker({
+      position: { lat: latitude, lng: longitude },
+      map: map,
+      title: "店舗の位置"
+    });
+  },
+
+  // 周辺検索の処理（Google Places APIを使用）
+  searchNearby: function(map) {
+    const service = new google.maps.places.PlacesService(map);
+    const request = {
+      location: map.getCenter(),
+      radius: '2000', // 検索範囲（メートル単位）
+      type: ['restaurant'],
+      keyword: 'ラーメン' // ラーメン店を検索
+    };
+
+    const self = this; // コンテキストを保持
+
+    service.nearbySearch(request, function(results, status) {
+      if (status === google.maps.places.PlacesServiceStatus.OK) {
+        results.forEach(function(place) {
+          self.createMarker(place, map); // ラーメン店にマーカーを追加
+        });
+      } else {
+        console.error("周辺検索に失敗しました: ", status);
+      }
+    });
+  },
+
+  // マップが静止した後にラーメン店を再検索
+  addIdleListener: function(map) {
+    const self = this;
+    map.addListener('idle', function() {
+      self.searchNearby(map);  // 地図の表示範囲が変更された後に検索
+    });
+  }
+};
+
+// ページロード後に適切なマップを初期化する
+document.addEventListener("turbo:load", function() {
+  // location-map の場合
+  const locationMapElement = document.getElementById("location-map");
+  if (locationMapElement) {
+    const latitude = parseFloat(locationMapElement.getAttribute("data-latitude"));
+    const longitude = parseFloat(locationMapElement.getAttribute("data-longitude"));
+    RamenMap.initLocationMap(latitude, longitude);
+  }
+
+  // user-map の場合
+  if (document.getElementById("user-map")) {
+    if (navigator.geolocation) {
+      navigator.geolocation.getCurrentPosition(
+        function(position) {
+          RamenMap.initUserMap(position);
+        },
+        function() {
+          RamenMap.initLocationMap(35.6895, 139.6917); // 東京をデフォルトに
+        }
+      );
+    } else {
+      RamenMap.initLocationMap(35.6895, 139.6917); // 東京をデフォルトに
     }
-  });
+  }
+
+  // store-map の場合
+  const storeMapElement = document.getElementById("store-map");
+  if (storeMapElement) {
+    const latitude = parseFloat(storeMapElement.getAttribute("data-latitude"));
+    const longitude = parseFloat(storeMapElement.getAttribute("data-longitude"));
+    RamenMap.initStoreMap(latitude, longitude);
+  }
+});
