+++ app/views/ramen_shops/_search_form.html.erb
@@ -27,20 +27,3 @@
   <%= f.submit "検索", class: "btn-primary search-btn" %>
   <ul id="autocomplete-list"></ul> 
 <% end %>
-
-<script>
-  document.addEventListener('turbo:load', function() {
-    // フォーム全体のリセット処理
-    const form = document.getElementById('search-form');
-    
-    document.getElementById('tab-name').addEventListener('click', function() {
-      form.reset();  // フォームをリセット
-    });
-    document.getElementById('tab-location').addEventListener('click', function() {
-      form.reset();
-    });
-    document.getElementById('tab-tag').addEventListener('click', function() {
-      form.reset();
-    });
-  });
-</script>
