+++ config/importmap.rb
@@ -5,5 +5,11 @@ pin "@hotwired/turbo-rails", to: "turbo.min.js"
 pin "@hotwired/stimulus", to: "stimulus.min.js"
 pin "@hotwired/stimulus-loading", to: "stimulus-loading.js"
 pin_all_from "app/javascript/controllers", under: "controllers"
-pin "avatar_preview", to: "avatar_preview.js"
 pin "videos", to: "videos.js"
+pin "search_form", to: "search_form.js"
+pin "my_page", to: "my_page.js"
+pin "map", to: "map.js"
+pin "location_information", to: "location_information.js"
+pin "comments", to: "comments.js"
+pin "avatar_preview", to: "avatar_preview.js"
+pin "autocomplete", to: "autocomplete.js"
